pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opensati"
version = "0.1.0"
description = "The Intelligent Mirror for Deep Work - Local AI copilot for mindful computing"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "OpenSati Team", email = "hello@opensati.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
]
keywords = ["productivity", "mindfulness", "focus", "ai", "local-first", "privacy"]

dependencies = [
    "customtkinter>=5.2.0",
    "pynput>=1.7.6",
    "mss>=9.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "ollama>=0.1.0",
    "pystray>=0.19.0",
    "screeninfo>=0.8.0",
]

[project.optional-dependencies]
audio = [
    "pyaudio>=0.2.14",
    "scipy>=1.11.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "opensati[audio,dev]",
]

[project.scripts]
opensati = "opensati.__main__:main"

[project.urls]
Homepage = "https://opensati.com"
Repository = "https://github.com/OpenSati-com/OpenSati"
Documentation = "https://github.com/OpenSati-com/OpenSati#readme"
Issues = "https://github.com/OpenSati-com/OpenSati/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
